requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.6.0
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langchain-community>=0.0.10
pandas>=2.2.0
numpy>=1.26.0
typing-extensions>=4.9.0
ipykernel